refactor: EditImageのロジック分離
